Start on data line.
Data line is probably not implemented properly? I don't know how to implement it, I don't know how data lines work. Need to do more research.

Deserialization is not complete, aligning the microsecond/cycle counter properly still needs to be implemented and then the rest of the deserialization will be complete. As long as the data line is implemented properly, that should be it for non-rumble commands.

Don't know how I'm going to fit rumble on a 24-pin board unless I knock the mode switch down to being a mode button/toggle.